pyproject.toml
[project]
name = "shellsage-cli"
version = "0.2.3"
description = "An AI-powered Linux command explainer (FAISS + Hugging Face)"
readme = "README.md"
authors = [{ name = "Nandita", email = "your.email@example.com" }]
license = "MIT"
requires-python = ">=3.9"
keywords = ["cli", "linux", "commands", "ai", "explainer", "terminal"]
dependencies = [
  "typer==0.12.3",
  "rich==13.7.1",
  "transformers==4.46.3",
  "huggingface_hub==0.24.6",
  "torch>=2.1.0",
  "faiss-cpu==1.12.0",
  "sentence-transformers==2.7.0",
  "numpy>=1.26.0",
  "tqdm==4.66.4",
  "PyPDF2==3.0.1",
  "pdfminer.six==20231228"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["shellsage*"]

[tool.setuptools.package-data]
shellsage = ["docs/*.txt", "embeddings/*"]

[project.scripts]
shellsage = "shellsage.cli:app"

